+++ src/assertions/openai.ts
@@ -1,5 +1,5 @@
 import type { OpenAiChatCompletionProvider } from '../providers/openai';
-import { validateFunctionCall } from '../providers/openaiUtil';
+import { validateFunctionCall } from '../providers/openai/util';
 import type { AssertionParams } from '../types';
 import type { GradingResult } from '../types';
 import { renderVarsInObject, maybeLoadFromExternalFile } from '../util';
